implemented new client and refactored to pep8 standard
