+++ application/config.py
@@ -5,4 +5,4 @@ SQLALCHEMY_DATABASE_URI = os.getenv('DEED_DATABASE_URI',
                                     'postgres:///deed_api')
 
 AKUMA_BASE_HOST = os.getenv('AKUMA_ADDRESS',
-                      'http://10.10.10.10:5055')
+                            'http://0.0.0.0:5055')
