+++ application/service_implimentation/akuma/__init__.py
@@ -2,4 +2,4 @@ from . import implementation, interface
 
 
 def make_akuma_client():
-    return interface.DeedApiInterface(implementation)
+    return interface.AkumaInterface(implementation)
