+++ application/service_implimentation/akuma/implementation.py
@@ -1,6 +1,13 @@
-class AkumaInterface(object):  # pragma: no cover
-    def __init__(self, implementation):
-        self.implementation = implementation
+import requests
+from application import config
+from flask.ext.api import status
 
-    def perform_check(self, payload):
-        return self.implementation.perform_check(payload)
+
+def perform_check(payload):  # pragma: no cover
+    data = None
+
+    resp = requests.post(config.AKUMA_BASE_HOST + "/", json=payload)
+
+    if resp.status_code == status.HTTP_200_OK:
+        data = resp.json()
+    return data
