+++ application/service_implimentation/akuma/interface.py
@@ -1,14 +1,7 @@
-import requests
-from application import config
-from flask.ext.api import status
+class AkumaInterface(object):  # pragma: no cover
 
+    def __init__(self, implementation):
+        self.implementation = implementation
 
-def perform_check(payload):  # pragma: no cover
-    data = None
-
-    resp = requests.get(config.AKUMA_BASE_HOST + "/", json=payload)
-
-    if resp.status_code == status.HTTP_200_OK:
-        data = resp.json()
-
-    return data
+    def perform_check(self, payload):
+        return self.implementation.perform_check(payload)
